# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Meeting Slot Service API
  version: 1.0.0
  description: |
    REST API for managing meeting slots and finding optimal meeting times for distributed teams.
    
    This service helps teams across multiple time zones find the best available meeting slots
    based on participant availability, working hours, and configurable preferences.
servers:
  - url: https://{api-gateway-id}.execute-api.{region}.amazonaws.com/v1
    description: AWS API Gateway
    variables:
      api-gateway-id:
        default: xxxxx
        description: API Gateway ID
      region:
        default: us-east-1
        description: AWS Region
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/v1/users:
    post:
      tags:
        - Users
      summary: Create a new user
      description: Creates a new user with name and email
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
            example:
              name: "John Doe"
              email: "john.doe@example.com"
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                success: true
                data:
                  id: "usr_abc123"
                  name: "John Doe"
                  email: "john.doe@example.com"
                  created_at: "2026-02-18T10:30:00Z"
                  updated_at: "2026-02-18T10:30:00Z"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "BAD_REQUEST"
                  message: "Invalid email format"

    get:
      tags:
        - Users
      summary: List all users
      description: Retrieves a list of users with optional pagination
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
              example:
                success: true
                data:
                  - id: "usr_abc123"
                    name: "John Doe"
                    email: "john.doe@example.com"
                    created_at: "2026-02-18T10:30:00Z"
                    updated_at: "2026-02-18T10:30:00Z"
                  - id: "usr_def456"
                    name: "Jane Smith"
                    email: "jane.smith@example.com"
                    created_at: "2026-02-18T11:00:00Z"
                    updated_at: "2026-02-18T11:00:00Z"
                pagination:
                  page: 1
                  limit: 20
                  total: 2
  /api/v1/users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: Retrieves a specific user by their unique identifier
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                success: true
                data:
                  id: "usr_abc123"
                  name: "John Doe"
                  email: "john.doe@example.com"
                  created_at: "2026-02-18T10:30:00Z"
                  updated_at: "2026-02-18T10:30:00Z"
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                error:
                  code: "NOT_FOUND"
                  message: "User not found"

    put:
      tags:
        - Users
      summary: Update user
      description: Updates an existing user's information
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
            example:
              name: "John Doe Updated"
              email: "john.updated@example.com"
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
              example:
                success: true
                data:
                  id: "usr_abc123"
                  name: "John Doe Updated"
                  email: "john.updated@example.com"
                  created_at: "2026-02-18T10:30:00Z"
                  updated_at: "2026-02-18T14:30:00Z"
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Users
      summary: Delete user
      description: Deletes a user by their unique identifier
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
    UserIdParam:
      name: id
      in: path
      required: true
      description: User unique identifier
      schema:
        type: string
        example: "usr_abc123"
...
    PageParam:
      name: page
      in: query
      description: Page number (default 1)
      schema:
        type: integer
        default: 1
        minimum: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page (default 20, max 100)
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
      example: 20
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: "NOT_FOUND"
        message:
          type: string
          description: Error message
          example: "Resource not found"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 100
...
    User:
      type: object
      properties:
        id:
          type: string
          description: Unique user identifier
          example: "usr_abc123"
        name:
          type: string
          description: User full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User email address
          example: "john.doe@example.com"
        created_at:
          type: string
          format: date-time
          description: User creation timestamp
          example: "2026-02-18T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: User last update timestamp
          example: "2026-02-18T10:30:00Z"
    CreateUserRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
          description: User full name
          example: "John Doe"
        email:
          type: string
          format: email
          description: User email address
          example: "john.doe@example.com"
    UpdateUserRequest:
      type: object
      properties:
        name:
          type: string
          description: User full name
          example: "John Doe Updated"
        email:
          type: string
          format: email
          description: User email address
          example: "john.updated@example.com"
    UserResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Event Schemas
...